==> 03 Scripts/1. Jupyter Fundamentals & Python Datatypes.ipynb <==
## Table of contents
## 01 Importing libraries
## 02 Python data types
## 02a Three ways of reaching 100 as a result
## 02b Coding two floating-point variables and dividing them by each other
## 02c Construct a short word 
## 02d Construct two short sentences

# 01 Importing libraries

import pandas as pd
import numpy as np
import os

# 02 Python data types

# 02a Three ways of reaching 100 as a result

x = 90

y = 10

x+y

x = 120

y = 20

x-y

x = 50

y = 50

x+y

# 02b Coding two floating-point variables and dividing them by each other

x = 5.534

y = 1.928

x/y

# 02c Construct a short word 

x = 'bill'

y = 'board'

x+y

# 02d Construct two short sentences

x = 'Hello '

y = 'world!'

print(x+y)

x = 'Python '

y = 'is a '

z = 'new language.'

print(x+y+z)
