==> 03 Scripts/2. Intro to Pandas.ipynb <==
## Table of contents
## 01 Importing libraries
## 02 Importing orders file
## 03 First 5 rows of dataframe
## 04 Importing products file
## 05 First 20 rows of dataframe
## 06 Last 35 rows of dataframe
## 07 Column names
## 08 Number of rows and columns
## 09 Descriptive statistics
## 10 Data Types

# 01 Importing libraries

import pandas as pd
import numpy as np
import os

# 02 Importing orders file

path = r'C:\Users\Eva\Documents\Instachart Basket Analysis'

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# 03 First 5 rows of dataframe

df.head()

# 04 Importing products file

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# 05 First 20 rows of dataframe

df_prods.head(20)

# 06 Last 35 rows of dataframe

df_prods.tail(35)

# 07 Column names

df_prods.columns

# 08 Number of rows and columns

df_prods.shape

# 09 Descriptive statistics

df_prods.describe()

# maximum value of the “aisle_id” column: 134.000000

# 10 Data Types

df_prods.dtypes

# data type of the “department_id” column: int64
