==> 03 Scripts/3. Data Import and Descriptive Analysis.ipynb <==
## Table of contents
## 01 Importing libraries
## 02 Importing data
## 03 Python Shortcut for Importing Data Files
## 04 Exploring Data with Pandas

# 01 Importing libraries

import pandas as pd
import numpy as np
import os

# 02 Importing data

df = pd.read_csv(r'C:\Users\Eva\Documents\Instachart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

# 03 Python Shortcut for Importing Data Files

path = r'C:\Users\Eva\Documents\Instachart Basket Analysis'

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# 04 Exploring Data with Pandas

df.head()

df.head(30)

df.tail()

df.shape

df.info()

df.dtypes

df.columns

df.describe()

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head()
