==> 03 Scripts/4. Data Wrangling and Subsetting.ipynb <==
## Table of contents
## 1 Importing libraries
## 2 Importing data files
## 3 Data Wrangling
###  3.1 Checking for missing values
### 3.2 Renaming columns
### 3.3 Changing a variable's data type
### 3.4 Transposing data
### 3.5 Data dictionary
### 3.6 Subsetting
## 4 Task
### Question 2: Another variable in df_ords that can be changed to string
### Question 3: Changing an unintuitive variable name
### Question 4: Busiest hour for placing orders
### Question 5: Meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary
### Question 6: A subset containing only information about breakfast item sales
### Question 7: All product information about items from the following departments: alcohol, deli, beverages, and meat/seafood
### Question 8: How many rows does the last dataframe you created have?
### Question 9: Extracting all information about user with user_id = 1
### Question 10: Basic stats about user1's behavior
## 5 Exporting dataframes

# 1 Importing libraries

import pandas as pd
import numpy as np
import os

# 2 Importing data files

path = r'C:\Users\Eva\Documents\Instachart Basket Analysis'

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# 3 Data Wrangling
##  3.1 Checking for missing values

df_ords['days_since_prior_order'].value_counts(dropna = False)

## 3.2 Renaming columns

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

## 3.3 Changing a variable's data type

df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

## 3.4 Transposing data

# Importing data set departments.csv

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

df_dep_t = df_dep.T

df_dep_t

# Take the first row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

new_header

# Take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Set the header row as the df header
df_dep_t_new.columns = new_header

df_dep_t_new

## 3.5 Data dictionary

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

## 3.6 Subsetting

df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# 4 Task

## Question 2: Another variable in df_ords that can be changed to string

df_ords.head()

df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

## Question 3: Changing an unintuitive variable name

## There are no more unintuitive names because I've already changed order_dow

## Question 4: Busiest hour for placing orders

df_ords['order_hour_of_day'].value_counts()

# 10 am is the busiest hour to place orders with orders being at 288418

## Question 5: Meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary

print(data_dict.get('4'))

# department_id = 4 in df_prods means that the department is "produce".

## Question 6: A subset containing only information about breakfast item sales

data_dict

# Breakfast items have the department number 14.

# Making a subset with only information about breakfast item sales
df_breakfast =  df_prods[df_prods['department_id']==14]

df_breakfast

## Question 7: All product information about items from the following departments: alcohol, deli, beverages, and meat/seafood

data_dict

# Alcohol, deli, beverages, and meat/seafood items have the department numbers 5, 20, 7, and 12 respectively.

# Making a subset with only information about alcohol, deli, beverages, and meat/seafood items
df_dinner_party =  df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_party

## Question 8: How many rows does the last dataframe you created have?

## It has 7650 rows (see above).

## Question 9: Extracting all information about user with user_id = 1

df_ords_user1 = df_ords[df_ords['user_id']=='1']

df_ords_user1

## Question 10: Basic stats about user1's behavior

df_ords_user1.describe()

# 5 Exporting dataframes

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
