==> 03 Scripts/5. Data Consistency Checks.ipynb <==
# Table of contents
# 1 Importing libraries
# 2 Importing data files
# 3 Mixed type data in test frame
# 4 Finding missing values in df_prods
# 5 Finding and addressing duplicates in df_prods
# 6 Exporting dataframes
# 7 Task
## Question 2: Run the df.describe() function on your df_ords dataframe
## Question 3: Check for mixed-type data in your df_ords dataframe
## Question 4: If you find mixed-type data, fix it. 
## Question 5: Check for missing values in df_ords dataframe
## Question 6: Address the missing values using an appropriate method
## Question 7: Check for duplicate values in your df_ords data
## Question 8: Address the duplicates using an appropriate method
## Question 9: Export your final, cleaned df_prods and df_ords data as “.csv” files 

# 1 Importing libraries

import pandas as pd
import numpy as np
import os

# 2 Importing data files

path = r'C:\Users\Eva\Documents\Instacart Basket Analysis'

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# 3 Mixed type data in test frame

## Create a dataframe with mixed type data
df_test = pd.DataFrame()

df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

## Checking for mixed type
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)
    

## Changing the data type to string
df_test['mix'] = df_test['mix'].astype('str')

# 4 Finding missing values in df_prods

## Finding the columns with missing values
df_prods.isnull().sum()

## Viewing missing values subset
df_nan = df_prods[df_prods['product_name'].isnull()==True]

df_nan

## There are 16 missing values found (also see above). 

## Rows and columns of the dataframe with missing values
df_prods.shape

## Rows and columns of the dataframe without missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# 5 Finding and addressing duplicates in df_prods

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

## 5 duplicates are found.

## Rows and columns of the dataframe with duplicates
df_prods_clean.shape

# Dropping the duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

## Rows and columns of the dataframe without duplicates
df_prods_clean_no_dups.shape

# 6 Exporting dataframes

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'), index = False)

# 7 Task

## Question 2: Run the df.describe() function on your df_ords dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.

df_ords.describe()

### I can drop the "unnamed: 0" columns because it is unnecessary. Apart from that, everything seems to be okay.

## Drop unnecessary column
df_ords = df_ords.drop(columns = ['Unnamed: 0'])

## Question 3: Check for mixed-type data in your df_ords dataframe

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)
    

### There is no mixed type data.

## Question 4: If you find mixed-type data, fix it. 

### There is no mixed type data.

## Question 5: Check for missing values in df_ords dataframe

## Finding missing values
df_ords.isnull().sum()

## Creating a subset of the missing values
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

## View missing values subset
df_ords_nan

## Viewing orders of user with user_id 1
df_ords[df_ords['user_id'] == 1]

## Viewing orders of user with user_id 2
df_ords[df_ords['user_id'] == 2]

## Viewing orders of user with user_id 3
df_ords[df_ords['user_id'] == 3]

### There are 206209 missing values in the column "days since prior order". As the examples from above show, missing values in this column mean that the customer ordered something for the very first time (order_number = 1), so no value could have been entered.

## Question 6: Address the missing values using an appropriate method

df_ords['days_since_prior_order'].fillna(0, inplace=True)

### I imputed 0 values for the NaN values, so that the values tell us that a customer has made its first order.

df_ords.head(30)

## Question 7: Check for duplicate values in your df_ords data

df_dups = df_ords[df_ords.duplicated()]

df_dups

### No duplicate data was found.

## Question 8: Address the duplicates using an appropriate method

### No duplicate data was found, so I don't need to do anything.

## Question 9: Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder and give them appropriate, succinct names.

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'), index = False)
