==> 03 Scripts/6. Combining and Exporting Data Part 1.ipynb <==
## Table of contents
## Question 1: Merge your prepared Instacart orders data with the new orders_products_prior dataframe 
## Question 2: Export of the merged file

# Question 1: Merge your prepared Instacart orders data with the new orders_products_prior dataframe 

# Importing libraries

import pandas as pd
import numpy as np
import os

# Importing data files

path = r'C:\Users\Eva\Documents\Instacart Basket Analysis'
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products_prior.csv'), index_col = False)
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

# Check the output

df_ords_prior.head()

df_ords.head()

df_ords_prior.shape

df_ords.shape

# Merging the dataframes

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check the output

df_merged_large.head()

df_merged_large.shape

# Question 2: Export of the merged file

# Export data to pkl

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
