==> 03 Scripts/6. Combining and Exporting Data Part 2.ipynb <==
## Table of contents
## Question 3: Import of pickle file and products file
## Question 4: Shape of the dataframes
## Question 5: Combine the orders_products_combined dataframe with your products data set
## Question 6: Confirm the results of the merge using the merge flag
## Question 7: Exporting the newly created dataframe

# Question 3: Import of pickle file and products file

# Importing libraries

import pandas as pd
import numpy as np
import os

# Set path

path = r'C:\Users\Eva\Documents\Instacart Basket Analysis'

# Importing data files
df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

df_prods.columns

# Check datatypes and memory usage for df_prods

df_prods.info()

# Change data types for df_prods

df_prods['product_id']=df_prods['product_id'].astype('int32')
df_prods['product_name'] = df_prods['product_name'].astype('object')
df_prods['aisle_id'] = df_prods['aisle_id'].astype('int8')
df_prods['department_id']=df_prods['department_id'].astype('int8')
df_prods['prices']=df_prods['prices'].astype('float16')

# Check output to see memory usage

df_prods.info()

# Check datatypes and memory usage for df_orders_products_combined 

df_orders_products_combined.info()

# Change types for df_orders_products_combined 

df_orders_products_combined['order_id']=df_orders_products_combined['order_id'].astype('int32')
df_orders_products_combined['user_id'] = df_orders_products_combined['user_id'].astype('int32')
df_orders_products_combined['order_number']=df_orders_products_combined['order_number'].astype('int8')
df_orders_products_combined['orders_day_of_week']=df_orders_products_combined['orders_day_of_week'].astype('int8')
df_orders_products_combined['order_hour_of_day']=df_orders_products_combined['order_hour_of_day'].astype('int8')
df_orders_products_combined['days_since_prior_order']=df_orders_products_combined['days_since_prior_order'].astype('float16')
df_orders_products_combined['product_id']=df_orders_products_combined['product_id'].astype('int32')
df_orders_products_combined['add_to_cart_order']=df_orders_products_combined['add_to_cart_order'].astype('int32')
df_orders_products_combined['reordered']=df_orders_products_combined['reordered'].astype('int8')

# Check memory usage for df_orders_products_combined 

df_orders_products_combined.info()

# Question 4: Shape of the dataframes

# Shape of order_products_combined dataframe

df_orders_products_combined.shape 

df_orders_products_combined.head()

# Shape of products dataframe 

df_prods.shape

df_prods.head()

# Question 5: Combine the orders_products_combined dataframe with your products data set

df_merged = df_orders_products_combined.merge(df_prods, on = 'product_id', indicator = True)

df_merged

# Question 6: Confirm the results of the merge using the merge flag.

# See above and also:
df_merged['_merge'].value_counts()

# Question 7: Exporting the newly created dataframe

df_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'merged_data.pkl'))
