==> 03 Scripts/7. Deriving New Variables.ipynb <==
## Table of contents
## 1 Importing libraries and data files
## 2 If-Statement with User-Defined Functions
## 3 Creating the variable "price label"
## 4 Creating the variable "busiest day"
## 5 Task
### Question 2: Creating the variable "busiest days"
### Question 3: Accuracy checks
### Question 4: Create a new column called “busiest_period_of_day” containing labels “Most orders,” “Average orders,” and “Fewest orders”
### Question 5: Frequencies of the new column/variable
### Question 6: Exporting files

## 1 Importing libraries and data files

# Importing libraries

import pandas as pd
import numpy as np
import os

# Importing data files

path = r'C:\Users\Eva\Documents\Instacart Basket Analysis'
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merged.pkl'))

## 2 If-Statement with User-Defined Functions

# Creating a subset

df_subset = df_ords_prods_merged[:1000000]

df_subset.shape

# Defining the price label function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Applying the price label function

df_subset['price_range'] = df.apply(price_label, axis=1)

## 3 Creating the variable "price label"

#If-statements with the loc-function

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_label'] = 'High-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_label'] = 'Mid-range product' 

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_label'] = 'Low-range product'

# Checking the frequencies

df_ords_prods_merged['price_label'].value_counts(dropna = False)

## 4 Creating the variable "busiest day"

# Checking the frequencies

df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)


# Creating an empty list

result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

df_ords_prods_merged['busiest_day'] = result

df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

## 5 Task 

### Question 2: Creating the variable "busiest days"

# Creating an empty list refering to the frequencies of "orders_day_of_week" from above: 
# Saturday (=0) and Sunday (=1) are the busiest days, Wednesday (=4) and Tuesday (=3) are the slowest days. 
# Every other day is regularly busy.

result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest days")
  elif value == 1:
    result.append("Busiest days")
  elif value == 4:
    result.append("Slowest days")
  elif value == 3:
    result.append("Slowest days")
  else:
    result.append("Regularly busy")

result

# Creating the new variable

df_ords_prods_merged['busiest_day_2'] = result

### Question 3: Accuracy checks

# Accuracy/frequency checks

df_ords_prods_merged['busiest_day_2'].value_counts(dropna = False)

 df_ords_prods_merged['busiest_day_2'].describe()

Nothing unusual spotted, everything seems to be as expected. "Regulaly busy" is the most frequent label (followed by "Busiest days" and "Slowest days"). Also, the values add up correctly to the overall count of 32404859 rows. 

### Question 4: Create a new column called “busiest_period_of_day” containing labels “Most orders,” “Average orders,” and “Fewest orders”

# Checking the frequencies

df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# Creating an empty list refering to the frequencies of "order_hour_of_day" from above:
# The top 8 hours of the day (in terms of highest frequencies of orders) get the label "Most orders", 
# the bottom 8 hours of the day get the label "Fewest orders" 
# and the 8 hours of the day in between the two get the label "Average orders".

result = []

for value in df_ords_prods_merged["order_hour_of_day"]:
  if value == 3:
    result.append("Fewest orders")
  elif value == 4:
    result.append("Fewest orders")  
  elif value == 2:
    result.append("Fewest orders")  
  elif value == 5:
    result.append("Fewest orders")  
  elif value == 1:
    result.append("Fewest orders")
  elif value == 0:
    result.append("Fewest orders")
  elif value == 6:
    result.append("Fewest orders")
  elif value == 23:
    result.append("Fewest orders")
  elif value == 9:
    result.append("Most orders")
  elif value == 16:
    result.append("Most orders")
  elif value == 12:
    result.append("Most orders")
  elif value == 13:
    result.append("Most orders")
  elif value == 15:
    result.append("Most orders")
  elif value == 14:
    result.append("Most orders")
  elif value == 11:
    result.append("Most orders")
  elif value == 10:
    result.append("Most orders")
  else:
    result.append("Average orders")

result

# Creating the new variable

df_ords_prods_merged['busiest_hour_of_day'] = result

### Question 5: Frequencies of the new column/variable

# Frequency checks

df_ords_prods_merged['busiest_hour_of_day'].value_counts(dropna = False)

df_ords_prods_merged['busiest_hour_of_day'].describe()

### Question 6: Exporting files

df_ords_prods_merged.to_csv(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merged_2.csv'))
