==> 03 Scripts/8. Grouping Data and Aggregating Variables.ipynb <==
## Table of contents
## 1 Importing libraries and data files
## 2 Grouping data
## 3 Aggregating data
## 4 Creating a loyality flag using loc
## 5 Task
### Question 2: Finding the aggregated mean of the “order_number” column grouped by “department_id”
### Question 3: Analyze the result. How do the results for the entire dataframe differ from those of the subset? 
### Question 4: Follow the instructions in the Exercise for creating a loyalty flag for existing customers using the transform() and loc() functions.
### Question 5: The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified
### Question 6: The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders 
### Question 7: In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers

## 1 Importing libraries and data files

# Importing libraries

import pandas as pd
import numpy as np
import os

# Importing data files

path = r'C:\Users\Eva\Documents\Instacart Basket Analysis'
df_ords_prods_merge = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merged_2.csv'))

# Creating a subset

df = df_ords_prods_merge[:1000000]

df.shape

df.head()

## 2 Grouping data

# Grouping the data by product_name

df.groupby('product_name')

## 3 Aggregating data

# Grouping the data by department_id and getting the mean values of order_number for each group

df.groupby('department_id').agg({'order_number': ['mean']})

# Another option for grouping the data by department_id and getting the mean values of order_number for each group

df.groupby('department_id')['order_number'].mean()

# Getting also the minimum and maximum of order_number for each department_id group

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

## 4 Creating a loyality flag using loc

df_ords_prods_merge['prices'] = df_ords_prods_merge['prices'].astype('float64')
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] == np.inf, 'prices'] = 99999

df_ords_prods_merge['max_order'] = df_ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

df_ords_prods_merge.head(15)

df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
df_ords_prods_merge.loc[(df_ords_prods_merge['max_order'] <= 40) & (df_ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

df_ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

df_ords_prods_merge.head(60)

## 5 Task

### Question 2: Finding the aggregated mean of the “order_number” column grouped by “department_id” 

df_ords_prods_merge.groupby('department_id')['order_number'].mean()

### Question 3: Analyze the result. How do the results for the entire dataframe differ from those of the subset? Include your comments in a markdown cell below the executed code.

As one can see, the order_number means for the respective departments are a little bit different (in most cases a little lower) for the entire dataset than for the subset.

### Question 4: Follow the instructions in the Exercise for creating a loyalty flag for existing customers using the transform() and loc() functions.

see above

### Question 5: The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

df_ords_prods_merge['prices'].describe()

df_ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

### Question 6: The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria:
If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.”
If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# Split the data into groups based on the 'user_id' column. Apply the transform() function on the 'prices' to get the avg price

df_ords_prods_merge['avg_price'] = df_ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

df_ords_prods_merge.loc[df_ords_prods_merge['avg_price'] >= 10, 'spending_flag'] = 'High spender'

df_ords_prods_merge.loc[df_ords_prods_merge['avg_price'] < 10, 'spending_flag'] = 'Low spender'

# View the output

df_ords_prods_merge.head()

### Question 7: In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:
If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# Split the df based on user_id, using transform()function on days_since_prior_order to get the median days_since_prior_order per user

df_ords_prods_merge['median_order_freq'] = df_ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

df_ords_prods_merge.loc[df_ords_prods_merge['median_order_freq'] > 20, 'ord_frequency_flag'] = 'Non-frequent customer'

df_ords_prods_merge.loc[df_ords_prods_merge['median_order_freq'] > 10 & (df_ords_prods_merge['median']<=20), 'ord_frequency_flag'] = 'Regular customer'

df_ords_prods_merge.loc[df_ords_prods_merge['median_order_freq'] <= 10, 'ord_frequency_flag'] = 'Frequent customer'

# View the output

df_ords_prods_merge.head()

# Dropping the column Unnamend: 0

df_ords_prods_merge.drop(columns = ['Unnamed: 0'], inplace = True)

df_ords_prods_merge.head()

# Exporting the datafile

df_ords_prods_merge.to_csv(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merged_4.csv'))
