==> 03 Scripts/9. Data Visualization Part 2.ipynb <==
## Table of contents
## Step 1-3: Importing data files and libraries
## Step 4: Wrangle the data
## Step 5: Complete the fundamental data quality and consistency checks 
## Step 6: Combine your customer data with the rest of your prepared Instacart data

## Step 1-3: Importing data files and libraries

# Importing libraries

import pandas as pd
import numpy as np
import os

# Importing customer data file

path = r'C:\Users\Lenovo\Documents\Instacart Basket Analysis'
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'),index_col = False)

## Step 4: Wrangle the data so that it follows consistent logic; for example, rename columns with illogical names and drop columns that don’t add anything to your analysis.

# Checking the column headers of the df

df_customers.head()

# Checking the dimensions of the df

df_customers.shape

# Renaming columns to be more more intuitive: changing Surnam to Last Name

df_customers.rename(columns = {'Surnam': 'Last Name'}, inplace = True)

# Renaming columns to be more intuitive: changing n_dependants to Number of Dependants

df_customers.rename(columns = {'n_dependants': 'Number of Dependants'}, inplace = True)

# Renaming columns to be more intuitive: changing fam_status to Marital Status

df_customers.rename(columns = {'fam_status': 'Marital Status'}, inplace = True)

# Checking the new column headers

df_customers.head()

I'm not sure which columns I will need for further analysis, so for the moment I will leave them as they are.

## Step 5: Complete the fundamental data quality and consistency checks you’ve learned throughout this Achievement; for example, check for and address missing values and duplicates, and convert any mixed-type data

# Checking data types

df_customers.dtypes

# While being a number, user_id contains only categorical information, so it could be 
# changed to a string variable.

df_customers['user_id'] = df_customers['user_id'].astype('str')

# Checking the output

df_customers['user_id'].dtypes

# Checking the descriptive statistics of the numerical columns

df_customers.describe()

# Checking for missing values

df_customers.isnull().sum()

There are 11259 missing values in the First Name column. I will fill the empty cells with NaN values since there is no way that I can imput or guess the missing values. 

# Filling the empty cells with NaN values

df_customers['First Name'].fillna('NaN', inplace=True)

# Checking for mixed-type data columns

for col in df_customers.columns.tolist():
  weird = (df_customers[[col]].applymap(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers[weird]) > 0:
    print (col)
  else:
    print ('No mixed dtypes', col)

There are no mixed-type data columns. 

# Exporting data file

df_customers.to_pickle(os.path.join(path, '02 Data','Original Data', 'customers_2.pkl'))

# Importing data file

path = r'C:\Users\Lenovo\Documents\Instacart Basket Analysis'
df_customers = pd.read_pickle(os.path.join(path, '02 Data', 'Original Data', 'customers_2.pkl'))

# Checking for duplicate records by creating a new dataframe that contains duplicate records

df_dups = df_customers[df_customers.duplicated()]

df_dups.head()

There are no duplicate records in the dataframe.

## Step 6: Combine your customer data with the rest of your prepared Instacart data. (Hint: Make sure the key columns are the same data type!)

# Importing the ords_prods_merged data file

df_ords_prods_merged = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merged_4.csv'), index_col = False)

# Checking the datafile

df_ords_prods_merged.shape

df_ords_prods_merged.columns

# Dropping an unnecessary column

df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['Unnamed: 0']) 

df_ords_prods_merged.shape

# Checking the datatypes

df_ords_prods_merged.dtypes

# Changing the datatype of the column user_id to string, so that it is the same datatype as user_id in the customers df

df_ords_prods_merged['user_id'] = df_ords_prods_merged['user_id'].astype('str')

df_ords_prods_merged['user_id'].dtypes

# Merging the two dfs with user_id as key variable

df_combined = df_ords_prods_merged.merge(df_customers, on = 'user_id')

# Cleaning the column headings for consistency

df_combined.rename(columns = {'First Name' : 'first_name', 'Last Name' : 'last_name', 'Gender' : 'gender', 'STATE' : 'state', 'Age' : 'age', 'Number of Dependants' : 'number_of_dependants', 'Marital Status' : 'marital_status'}, inplace = True)

df_combined.head()

# Exporting the datafile

df_combined.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'all_combined.pkl'))
